+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.3.0] - 2023-02-14
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Added `Footer component`.
+- [Marek Kurańda](https://github.com/mjkuranda): Added `Header component`.
 - [Marek Kurańda](https://github.com/mjkuranda): Added `result` subpage.
 
 ## [0.2.0] - 2023-02-11