+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.0] - 2023-02-17
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Added basic `search` subpage containing some filters.
-- [Marek Kurańda](https://github.com/mjkuranda): Added `Ingredient Component`.
+- [Marek Kurańda](https://github.com/mjkuranda): Added many components for filters in Search subsite.
+- [Marek Kurańda](https://github.com/mjkuranda): Building full params query for `ings` and `types`;
 
 ## [0.3.0] - 2023-02-14
 ### Added