+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2023-03-11
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Added meal results and submit button.
+
 ## [0.4.0] - 2023-02-17
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Added basic `search` subpage containing some filters.